# backend/backend/docs/swagger/messenger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Suda Messenger: authentication, chats, messaging, users, search.",
        "title": "Suda Messenger Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/channels/by-username/{username}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns channel_id by channel @username. The \"@\" prefix is trimmed automatically.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "summary": "Resolve channel by username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel username (with or without @)",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "channel_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/channels/comments/{comment_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Only the comment author can edit their comment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "summary": "Edit a comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New content",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_channel.EditCommentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Comment author or channel OWNER/ADMIN can delete.",
                "tags": [
                    "Channels"
                ],
                "summary": "Delete a comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/channels/{id}/apps": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "summary": "Get apps linked to a channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_channel.ChannelAppLink"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Channel OWNER/ADMIN only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "summary": "Link an app to a channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "App ID to link",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_channel.LinkAppReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/channels/{id}/unsubscribe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "summary": "Unsubscribe from a channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_channel.EditCommentReq": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 4000,
                    "minLength": 1
                }
            }
        },
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_channel.LinkAppReq": {
            "type": "object",
            "required": [
                "app_id"
            ],
            "properties": {
                "app_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}